requirements.txt
fastapi==0.104.1
sqlalchemy==2.0.25
jinja2==3.1.2
python-multipart==0.0.6
bcrypt==4.1.2
uvicorn[standard]==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
itsdangerous==2.1.2